migrate middleware tests
