+++ test/e2e/middleware-base-path/app/next.config.js
