+++ test/e2e/middleware-base-path/app/pages/index.js
