+++ test/e2e/middleware-general/test/index.test.ts
@@ -1,66 +1,64 @@
 /* eslint-env jest */
 
-import { join } from 'path'
 import fs from 'fs-extra'
+import { join } from 'path'
 import webdriver from 'next-webdriver'
-import {
-  check,
-  fetchViaHTTP,
-  findPort,
-  killApp,
-  launchApp,
-  nextBuild,
-  nextStart,
-  waitFor,
-} from 'next-test-utils'
-
-jest.setTimeout(1000 * 60 * 2)
+import { NextInstance } from 'test/lib/next-modes/base'
+import { check, fetchViaHTTP, renderViaHTTP, waitFor } from 'next-test-utils'
+import { createNext, FileRef } from 'e2e-utils'
+import escapeStringRegexp from 'escape-string-regexp'
 
 const middlewareWarning = 'using beta Middleware (not covered by semver)'
 const urlsError = 'Please use only absolute URLs'
-const context = {
-  appDir: join(__dirname, '../'),
-  buildLogs: { output: '', stdout: '', stderr: '' },
-  logs: { output: '', stdout: '', stderr: '' },
-}
 
 describe('Middleware Runtime', () => {
-  describe('dev mode', () => {
-    afterAll(() => killApp(context.app))
-    beforeAll(async () => {
-      context.dev = true
-      context.appPort = await findPort()
-      context.buildId = 'development'
-      context.app = await launchApp(context.appDir, context.appPort, {
-        env: {
-          MIDDLEWARE_TEST: 'asdf',
-          NEXT_RUNTIME: 'edge',
+  let next: NextInstance
+  let locale = ''
+
+  afterAll(() => next.destroy())
+  beforeAll(async () => {
+    next = await createNext({
+      files: {
+        'next.config.js': new FileRef(join(__dirname, '../app/next.config.js')),
+        'middleware.js': new FileRef(join(__dirname, '../app/middleware.js')),
+        pages: new FileRef(join(__dirname, '../app/pages')),
+        'shared-package': new FileRef(
+          join(__dirname, '../app/node_modules/shared-package')
+        ),
+      },
+      packageJson: {
+        scripts: {
+          setup: `cp -r ./shared-package ./node_modules`,
+          build: 'yarn setup && next build',
+          dev: 'yarn setup && next dev',
+          start: 'next start',
         },
-        onStdout(msg) {
-          context.logs.output += msg
-          context.logs.stdout += msg
-        },
-        onStderr(msg) {
-          context.logs.output += msg
-          context.logs.stderr += msg
-        },
-      })
+      },
+      startCommand: (global as any).isNextDev ? 'yarn dev' : 'yarn start',
+      buildCommand: 'yarn build',
+      env: {
+        MIDDLEWARE_TEST: 'asdf',
+        NEXT_RUNTIME: 'edge',
+      },
     })
+  })
 
-    tests(context)
-
-    // This test has to be after something has been executed with middleware
-    it('should have showed warning for middleware usage', () => {
-      expect(context.logs.output).toContain(middlewareWarning)
+  if ((global as any).isNextDev) {
+    it('should have showed warning for middleware usage', async () => {
+      await renderViaHTTP(next.url, '/')
+      await check(
+        () => next.cliOutput,
+        new RegExp(escapeStringRegexp(middlewareWarning))
+      )
     })
 
     it('refreshes the page when middleware changes ', async () => {
-      const browser = await webdriver(context.appPort, `/about`)
+      const browser = await webdriver(next.url, `/about`)
       await browser.eval('window.didrefresh = "hello"')
       const text = await browser.elementByCss('h1').text()
       expect(text).toEqual('AboutA')
 
-      const middlewarePath = join(context.appDir, '/middleware.js')
+      const middlewarePath = join(next.testDir, '/middleware.js')
       const originalContent = fs.readFileSync(middlewarePath, 'utf-8')
       const editedContent = originalContent.replace('/about/a', '/about/b')
 
@@ -75,48 +73,12 @@ describe('Middleware Runtime', () => {
         await browser.close()
       }
     })
-  })
+  }
 
-  describe('production mode', () => {
-    afterAll(() => killApp(context.app))
-    beforeAll(async () => {
-      const build = await nextBuild(context.appDir, undefined, {
-        stderr: true,
-        stdout: true,
-      })
-
-      context.buildId = await fs.readFile(
-        join(context.appDir, '.next/BUILD_ID'),
-        'utf8'
-      )
-
-      context.buildLogs = {
-        output: build.stdout + build.stderr,
-        stderr: build.stderr,
-        stdout: build.stdout,
-      }
-      context.dev = false
-
-      context.appPort = await findPort()
-      context.app = await nextStart(context.appDir, context.appPort, {
-        env: {
-          MIDDLEWARE_TEST: 'asdf',
-          NEXT_RUNTIME: 'edge',
-        },
-        onStdout(msg) {
-          context.logs.output += msg
-          context.logs.stdout += msg
-        },
-        onStderr(msg) {
-          context.logs.output += msg
-          context.logs.stderr += msg
-        },
-      })
-    })
-
+  if ((global as any).isNextStart) {
     it('should have valid middleware field in manifest', async () => {
       const manifest = await fs.readJSON(
-        join(context.appDir, '.next/server/middleware-manifest.json')
+        join(next.testDir, '.next/server/middleware-manifest.json')
       )
       expect(manifest.middleware).toEqual({
         '/': {
@@ -131,16 +93,16 @@ describe('Middleware Runtime', () => {
     })
 
     it('should have middleware warning during build', () => {
-      expect(context.buildLogs.output).toContain(middlewareWarning)
+      expect(next.cliOutput).toContain(middlewareWarning)
     })
 
     it('should have middleware warning during start', () => {
-      expect(context.logs.output).toContain(middlewareWarning)
+      expect(next.cliOutput).toContain(middlewareWarning)
     })
 
     it('should have correct files in manifest', async () => {
       const manifest = await fs.readJSON(
-        join(context.appDir, '.next/server/middleware-manifest.json')
+        join(next.testDir, '.next/server/middleware-manifest.json')
       )
       for (const key of Object.keys(manifest.middleware)) {
         const middleware = manifest.middleware[key]
@@ -152,15 +114,12 @@ describe('Middleware Runtime', () => {
         )
       }
     })
+  }
 
-    tests(context)
-  })
-})
-
-function tests(context, locale = '') {
+  // TODO: re-enable after fixing server-side resolving priority
   it('should redirect the same for direct visit and client-transition', async () => {
     const res = await fetchViaHTTP(
-      context.appPort,
+      next.url,
       `${locale}/redirect-1`,
       undefined,
       {
@@ -172,7 +131,7 @@ function tests(context, locale = '') {
       '/somewhere-else'
     )
 
-    const browser = await webdriver(context.appPort, `${locale}/`)
+    const browser = await webdriver(next.url, `${locale}/`)
     await browser.eval(`next.router.push('/redirect-1')`)
     await check(async () => {
       const pathname = await browser.eval('location.pathname')
@@ -180,12 +139,13 @@ function tests(context, locale = '') {
     }, 'success')
   })
 
+  // TODO: re-enable after fixing server-side resolving priority
   it('should rewrite the same for direct visit and client-transition', async () => {
-    const res = await fetchViaHTTP(context.appPort, `${locale}/rewrite-1`)
+    const res = await fetchViaHTTP(next.url, `${locale}/rewrite-1`)
     expect(res.status).toBe(200)
     expect(await res.text()).toContain('Hello World')
 
-    const browser = await webdriver(context.appPort, `${locale}/`)
+    const browser = await webdriver(next.url, `${locale}/`)
     await browser.eval(`next.router.push('/rewrite-1')`)
     await check(async () => {
       const content = await browser.eval('document.documentElement.innerHTML')
@@ -194,11 +154,11 @@ function tests(context, locale = '') {
   })
 
   it('should rewrite correctly for non-SSG/SSP page', async () => {
-    const res = await fetchViaHTTP(context.appPort, `${locale}/rewrite-2`)
+    const res = await fetchViaHTTP(next.url, `${locale}/rewrite-2`)
     expect(res.status).toBe(200)
     expect(await res.text()).toContain('AboutA')
 
-    const browser = await webdriver(context.appPort, `${locale}/`)
+    const browser = await webdriver(next.url, `${locale}/`)
     await browser.eval(`next.router.push('/rewrite-2')`)
     await check(async () => {
       const content = await browser.eval('document.documentElement.innerHTML')
@@ -207,61 +167,74 @@ function tests(context, locale = '') {
   })
 
   it('should respond with 400 on decode failure', async () => {
-    const res = await fetchViaHTTP(context.appPort, `${locale}/%2`)
+    const res = await fetchViaHTTP(next.url, `${locale}/%2`)
     expect(res.status).toBe(400)
 
-    if (!context.dev) {
+    if ((global as any).isNextStart) {
       expect(await res.text()).toContain('Bad Request')
     }
   })
 
-  it('should set fetch user agent correctly', async () => {
-    const res = await fetchViaHTTP(
-      context.appPort,
-      `${locale}/fetch-user-agent-default`
-    )
-    expect(readMiddlewareJSON(res).headers['user-agent']).toBe(
-      'Next.js Middleware'
-    )
+  if (!(global as any).isNextDeploy) {
+    // user agent differs on Vercel
+    it('should set fetch user agent correctly', async () => {
+      const res = await fetchViaHTTP(
+        next.url,
+        `${locale}/fetch-user-agent-default`
+      )
 
-    const res2 = await fetchViaHTTP(
-      context.appPort,
-      `${locale}/fetch-user-agent-crypto`
-    )
-    expect(readMiddlewareJSON(res2).headers['user-agent']).toBe('custom-agent')
-  })
+      expect(readMiddlewareJSON(res).headers['user-agent']).toBe(
+        'Next.js Middleware'
+      )
+
+      const res2 = await fetchViaHTTP(
+        next.url,
+        `${locale}/fetch-user-agent-crypto`
+      )
+      expect(readMiddlewareJSON(res2).headers['user-agent']).toBe(
+        'custom-agent'
+      )
+    })
+  }
 
   it('should contain process polyfill', async () => {
-    const res = await fetchViaHTTP(context.appPort, `/global`)
+    const res = await fetchViaHTTP(next.url, `/global`)
     expect(readMiddlewareJSON(res)).toEqual({
       process: {
         env: {
           MIDDLEWARE_TEST: 'asdf',
-          NEXT_RUNTIME: 'edge',
+          ...((global as any).isNextDeploy
+            ? {}
+            : {
+                NEXT_RUNTIME: 'edge',
+              }),
         },
       },
     })
   })
 
   it(`should contain \`globalThis\``, async () => {
-    const res = await fetchViaHTTP(context.appPort, '/globalthis')
+    const res = await fetchViaHTTP(next.url, '/globalthis')
     expect(readMiddlewareJSON(res).length > 0).toBe(true)
   })
 
   it(`should contain crypto APIs`, async () => {
-    const res = await fetchViaHTTP(context.appPort, '/webcrypto')
+    const res = await fetchViaHTTP(next.url, '/webcrypto')
     expect('error' in readMiddlewareJSON(res)).toBe(false)
   })
 
-  it(`should accept a URL instance for fetch`, async () => {
-    const response = await fetchViaHTTP(context.appPort, '/fetch-url')
-    const { error } = readMiddlewareJSON(response)
-    expect(error).toBeTruthy()
-    expect(error.message).not.toContain("Failed to construct 'URL'")
-  })
+  if (!(global as any).isNextDeploy) {
+    it(`should accept a URL instance for fetch`, async () => {
+      const response = await fetchViaHTTP(next.url, '/fetch-url')
+      // TODO: why is an error expected here if it should work?
+      const { error } = readMiddlewareJSON(response)
+      expect(error).toBeTruthy()
+      expect(error.message).not.toContain("Failed to construct 'URL'")
+    })
+  }
 
   it(`should allow to abort a fetch request`, async () => {
-    const response = await fetchViaHTTP(context.appPort, '/abort-controller')
+    const response = await fetchViaHTTP(next.url, '/abort-controller')
     const payload = readMiddlewareJSON(response)
     expect('error' in payload).toBe(true)
     expect(payload.error.name).toBe('AbortError')
@@ -269,9 +242,9 @@ function tests(context, locale = '') {
   })
 
   it(`should validate & parse request url from any route`, async () => {
-    const res = await fetchViaHTTP(context.appPort, `${locale}/static`)
+    const res = await fetchViaHTTP(next.url, `${locale}/static`)
 
-    expect(res.headers.get('req-url-basepath')).toBe('')
+    expect(res.headers.get('req-url-basepath')).toBeFalsy()
     expect(res.headers.get('req-url-pathname')).toBe('/static')
 
     const { pathname, params } = JSON.parse(res.headers.get('req-url-params'))
@@ -285,9 +258,9 @@ function tests(context, locale = '') {
   })
 
   it(`should validate & parse request url from a dynamic route with params`, async () => {
-    const res = await fetchViaHTTP(context.appPort, `/fr/1`)
+    const res = await fetchViaHTTP(next.url, `/fr/1`)
 
-    expect(res.headers.get('req-url-basepath')).toBe('')
+    expect(res.headers.get('req-url-basepath')).toBeFalsy()
     expect(res.headers.get('req-url-pathname')).toBe('/1')
 
     const { pathname, params } = JSON.parse(res.headers.get('req-url-params'))
@@ -299,8 +272,8 @@ function tests(context, locale = '') {
   })
 
   it(`should validate & parse request url from a dynamic route with params and no query`, async () => {
-    const res = await fetchViaHTTP(context.appPort, `/fr/abc123`)
-    expect(res.headers.get('req-url-basepath')).toBe('')
+    const res = await fetchViaHTTP(next.url, `/fr/abc123`)
+    expect(res.headers.get('req-url-basepath')).toBeFalsy()
 
     const { pathname, params } = JSON.parse(res.headers.get('req-url-params'))
     expect(pathname).toBe('/:locale/:id')
@@ -311,8 +284,8 @@ function tests(context, locale = '') {
   })
 
   it(`should validate & parse request url from a dynamic route with params and query`, async () => {
-    const res = await fetchViaHTTP(context.appPort, `/abc123?foo=bar`)
-    expect(res.headers.get('req-url-basepath')).toBe('')
+    const res = await fetchViaHTTP(next.url, `/abc123?foo=bar`)
+    expect(res.headers.get('req-url-basepath')).toBeFalsy()
 
     const { pathname, params } = JSON.parse(res.headers.get('req-url-params'))
 
@@ -324,57 +297,45 @@ function tests(context, locale = '') {
   })
 
   it('should throw when using URL with a relative URL', async () => {
-    const res = await fetchViaHTTP(context.appPort, `/url/relative-url`)
+    const res = await fetchViaHTTP(next.url, `/url/relative-url`)
     expect(readMiddlewareError(res)).toContain('Invalid URL')
   })
 
-  it('should throw when using Request with a relative URL', async () => {
-    const response = await fetchViaHTTP(
-      context.appPort,
-      `/url/relative-request`
-    )
-    expect(readMiddlewareError(response)).toContain(urlsError)
-  })
-
   it('should throw when using NextRequest with a relative URL', async () => {
-    const response = await fetchViaHTTP(
-      context.appPort,
-      `/url/relative-next-request`
-    )
+    const response = await fetchViaHTTP(next.url, `/url/relative-next-request`)
     expect(readMiddlewareError(response)).toContain(urlsError)
   })
 
-  it('should warn when using Response.redirect with a relative URL', async () => {
-    const response = await fetchViaHTTP(
-      context.appPort,
-      `/url/relative-redirect`
-    )
-    expect(readMiddlewareError(response)).toContain(urlsError)
-  })
+  if (!(global as any).isNextDeploy) {
+    // these errors differ on Vercel
+    it('should throw when using Request with a relative URL', async () => {
+      const response = await fetchViaHTTP(next.url, `/url/relative-request`)
+      expect(readMiddlewareError(response)).toContain(urlsError)
+    })
+
+    it('should warn when using Response.redirect with a relative URL', async () => {
+      const response = await fetchViaHTTP(next.url, `/url/relative-redirect`)
+      expect(readMiddlewareError(response)).toContain(urlsError)
+    })
+  }
 
   it('should warn when using NextResponse.redirect with a relative URL', async () => {
-    const response = await fetchViaHTTP(
-      context.appPort,
-      `/url/relative-next-redirect`
-    )
+    const response = await fetchViaHTTP(next.url, `/url/relative-next-redirect`)
     expect(readMiddlewareError(response)).toContain(urlsError)
   })
 
   it('should throw when using NextResponse.rewrite with a relative URL', async () => {
-    const response = await fetchViaHTTP(
-      context.appPort,
-      `/url/relative-next-rewrite`
-    )
+    const response = await fetchViaHTTP(next.url, `/url/relative-next-rewrite`)
     expect(readMiddlewareError(response)).toContain(urlsError)
   })
 
   it('should trigger middleware for data requests', async () => {
-    const browser = await webdriver(context.appPort, `/ssr-page`)
+    const browser = await webdriver(next.url, `/ssr-page`)
     const text = await browser.elementByCss('h1').text()
     expect(text).toEqual('Bye Cruel World')
     const res = await fetchViaHTTP(
-      context.appPort,
-      `/_next/data/${context.buildId}/en/ssr-page.json`
+      next.url,
+      `/_next/data/${next.buildId}/en/ssr-page.json`
     )
     const json = await res.json()
     expect(json.pageProps.message).toEqual('Bye Cruel World')
@@ -382,27 +343,27 @@ function tests(context, locale = '') {
 
   it('should normalize data requests into page requests', async () => {
     const res = await fetchViaHTTP(
-      context.appPort,
-      `/_next/data/${context.buildId}/en/send-url.json`
+      next.url,
+      `/_next/data/${next.buildId}/en/send-url.json`
     )
     expect(res.headers.get('req-url-path')).toEqual('/send-url')
   })
 
   it('should keep non data requests in their original shape', async () => {
     const res = await fetchViaHTTP(
-      context.appPort,
-      `/_next/static/${context.buildId}/_devMiddlewareManifest.json?foo=1`
+      next.url,
+      `/_next/static/${next.buildId}/_devMiddlewareManifest.json?foo=1`
     )
     expect(res.headers.get('req-url-path')).toEqual(
-      `/_next/static/${context.buildId}/_devMiddlewareManifest.json?foo=1`
+      `/_next/static/${next.buildId}/_devMiddlewareManifest.json?foo=1`
     )
   })
 
   it('should add a rewrite header on data requests for rewrites', async () => {
-    const res = await fetchViaHTTP(context.appPort, `/ssr-page`)
+    const res = await fetchViaHTTP(next.url, `/ssr-page`)
     const dataRes = await fetchViaHTTP(
-      context.appPort,
-      `/_next/data/${context.buildId}/en/ssr-page.json`
+      next.url,
+      `/_next/data/${next.buildId}/en/ssr-page.json`
     )
     const json = await dataRes.json()
     expect(json.pageProps.message).toEqual('Bye Cruel World')
@@ -413,13 +374,13 @@ function tests(context, locale = '') {
   })
 
   it(`hard-navigates when the data request failed`, async () => {
-    const browser = await webdriver(context.appPort, `/error`)
+    const browser = await webdriver(next.url, `/error`)
     await browser.eval('window.__SAME_PAGE = true')
     await browser.elementByCss('#throw-on-data').click()
     await browser.waitForElementByCss('.refreshed')
     expect(await browser.eval('window.__SAME_PAGE')).toBeUndefined()
   })
-}
+})
 
 function readMiddlewareJSON(response) {
   return JSON.parse(response.headers.get('data'))
