+++ test/e2e/middleware-redirects/app/middleware.js
@@ -1,5 +1,12 @@
+import { NextResponse } from 'next/server'
+
 export async function middleware(request) {
   const url = request.nextUrl
+
+  // this is needed for tests to get the BUILD_ID
+  if (url.pathname.startsWith('/_next/static/__BUILD_ID')) {
+    return NextResponse.next()
+  }
 
   if (url.pathname === '/old-home') {
     if (url.searchParams.get('override') === 'external') {
