+++ test/e2e/middleware-redirects/app/next.config.js
