+++ test/e2e/middleware-redirects/app/pages/index.js
