+++ test/e2e/middleware-redirects/app/pages/new-home.js
