+++ test/e2e/middleware-responses/app/pages/index.js
