+++ test/e2e/middleware-rewrites/app/next.config.js
