+++ test/e2e/middleware-rewrites/app/pages/[param].js
