+++ test/e2e/middleware-rewrites/app/pages/i18n.js
