+++ test/e2e/middleware-rewrites/app/pages/index.js
