+++ test/lib/next-modes/next-deploy.ts
@@ -65,16 +65,26 @@ export class NextDeployInstance extends NextInstance {
     }
     require('console').log(`Deploying project at ${this.testDir}`)
 
+    const additionalEnv = []
+
+    for (const key of Object.keys(this.env || {})) {
+      additionalEnv.push('--build-env')
+      additionalEnv.push(`${key}=${this.env[key]}`)
+      additionalEnv.push('--env')
+      additionalEnv.push(`${key}=${this.env[key]}`)
+    }
+
     const deployRes = await execa(
       'vercel',
       [
         'deploy',
         '--build-env',
         'NEXT_PRIVATE_TEST_MODE=1',
-        '--build-env',
-        'FORCE_RUNTIME_TAG=canary',
+        // '--build-env',
+        // 'FORCE_RUNTIME_TAG=canary',
         '--build-env',
         'NEXT_TELEMETRY_DISABLED=1',
+        ...additionalEnv,
         '--force',
         ...vercelFlags,
       ],
