+++ test/lib/next-modes/next-start.ts
@@ -46,6 +46,7 @@ export class NextStartInstance extends NextInstance {
       shell: false,
       env: {
         ...process.env,
+        ...this.env,
         NODE_ENV: '' as any,
         __NEXT_TEST_MODE: '1',
         __NEXT_RAND_PORT: '1',
